+++ scripts/applications/medkit-effect.js
@@ -122,6 +122,11 @@ export class EffectMedkit extends HandlebarsApplicationMixin(ApplicationV2) {
                             label: 'CHRISPREMADES.Medkit.Effect.SpecialDuration.EndOfWorkflow',
                             value: 'endOfWorkflow',
                             isSelected: effect.flags['chris-premades']?.specialDuration?.includes('endOfWorkflow')
+                        },
+                        {
+                            label: 'CHRISPREMADES.Medkit.Effect.SpecialDuration.MoveFinished',
+                            value: 'moveFinished',
+                            isSelected: effect.flags['chris-premades']?.specialDuration?.includes('moveFinished')
                         }
                     ].concat(CONFIG.statusEffects.map(i => ({
                         label: i.name,