+++ scripts/macros.js
@@ -281,6 +281,17 @@ export {guidedStrike} from './macros/2024/classFeatures/cleric/war/guidedStrike.
 export {warPriest} from './macros/2024/classFeatures/cleric/war/warPriest.js';
 export {warGodsBlessing, warGodsBlessingCast} from './macros/2024/classFeatures/cleric/war/warGodsBlessing.js';
 export {avatarOfBattle} from './macros/2024/classFeatures/cleric/war/avatarOfBattle.js';
+//Ranger
+export {favoredEnemy} from './macros/2024/classFeatures/ranger/favoredEnemy.js';
+export {tireless} from './macros/2024/classFeatures/ranger/tireless.js';
+export {relentlessHunter} from './macros/2024/classFeatures/ranger/relentlessHunter.js';
+export {naturesVeil} from './macros/2024/classFeatures/ranger/naturesVeil.js';
+export {preciseHunter} from './macros/2024/classFeatures/ranger/preciseHunter.js';
+export {foeSlayer} from './macros/2024/classFeatures/ranger/foeSlayer.js';
+//Beast Master
+export {primalCompanion, bestialFury} from './macros/2024/classFeatures/ranger/beastMaster/primalCompanion.js';
+export {exceptionalTraining} from './macros/2024/classFeatures/ranger/beastMaster/exceptionalTraining.js';
+export {shareSpells} from './macros/2024/classFeatures/ranger/beastMaster/shareSpells.js';
 //Rogue
 export {sneakAttack} from './macros/2024/classFeatures/rogue/sneakAttack.js';
 export {cunningAction} from './macros/2024/classFeatures/rogue/cunningAction.js';