+++ scripts/macros/2014/classFeatures/illrigger/painkiller/grandStrategist.js
@@ -1,0 +1,26 @@
+import {itemUtils} from '../../../../../utils.js';
+async function added({trigger: {entity: item}}) {
+    await itemUtils.correctActivityItemConsumption(item, ['use'], 'invokeHell');
+}
+export let grandStrategist = {
+    name: 'Invoke Hell: Grand Strategist',
+    version: '1.3.78',
+    rules: 'legacy',
+    item: [
+        {
+            pass: 'created',
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
+        }
+    ]
+};