+++ scripts/macros/2024/classFeatures/ranger/beastMaster/exceptionalTraining.js
@@ -1,0 +1,5 @@
+export let exceptionalTraining  = {
+    name: 'Exceptional Training',
+    version: '1.3.79',
+    rules: 'modern'
+};