+++ scripts/macros/2024/classFeatures/ranger/beastMaster/primalCompanion.js
@@ -1,5 +1,5 @@
 import {Summons} from '../../../../../lib/summons.js';
-import {activityUtils, actorUtils, compendiumUtils, constants, dialogUtils, errors, genericUtils, itemUtils} from '../../../../../utils.js';
+import {activityUtils, actorUtils, combatUtils, compendiumUtils, constants, dialogUtils, effectUtils, errors, genericUtils, itemUtils, workflowUtils} from '../../../../../utils.js';
 
 async function use({workflow}) {
     let activityIdentifier = activityUtils.getIdentifier(workflow.activity);
@@ -8,13 +8,29 @@ async function use({workflow}) {
     let classLevel = workflow.actor.classes?.ranger?.system?.levels;
     if (!classLevel) return;
     let primalBondFeatureData = await Summons.getSummonItem('Primal Bond', {}, workflow.item, {translate: 'CHRISPREMADES.Macros.PrimalCompanion.PrimalBond', identifier: 'primalCompanionPrimalBond'});
+    let dashData = await compendiumUtils.getItemFromCompendium(constants.packs.actions, 'Dash', {object: true, getDescription: true, translate: 'CHRISPREMADES.Macros.Actions.Dash', identifier: 'primalCompanionDash'});
+    let disengageData = await compendiumUtils.getItemFromCompendium(constants.packs.actions, 'Disengage', {object: true, getDescription: true, translate: 'CHRISPREMADES.Macros.Actions.Disengage', identifier: 'primalCompanionDisengage'});
     let dodgeData = await compendiumUtils.getItemFromCompendium(constants.packs.actions, 'Dodge', {object: true, getDescription: true, translate: 'CHRISPREMADES.Macros.Actions.Dodge', identifier: 'primalCompanionDodge'});
-    if (!primalBondFeatureData || !dodgeData) {
+    let helpData = await compendiumUtils.getItemFromCompendium(constants.packs.actions, 'Help', {object: true, getDescription: true, translate: 'CHRISPREMADES.Macros.Actions.Help', identifier: 'primalCompanionHelp'});
+    if (!primalBondFeatureData || !dashData || !disengageData || !dodgeData || !helpData) {
         errors.missingPackItem();
         return;
     }
+    let itemsToAdd = [primalBondFeatureData];
     let commandFeature = activityUtils.getActivityByIdentifier(workflow.item, 'primalCompanionCommand', {strict: true});
     if (!commandFeature) return;
+    let exceptionalTraining = itemUtils.getItemByIdentifier(workflow.actor, 'exceptionalTraining');
+    if (exceptionalTraining) {
+        let genericActions = [dashData, disengageData, dodgeData, helpData];
+        genericActions.forEach(i => {
+            let genericActivity = Object.entries(i.system.activities)[0][1]
+            genericActivity.activation.type = 'bonus';
+            itemsToAdd.push(i);
+        })
+    }
+    else {
+        itemsToAdd.push(dodgeData);
+    }
     let creatureType = activityIdentifier.slice(15).toLowerCase();
     let hpValue = 5 + (classLevel * 5);
     let name = itemUtils.getConfig(workflow.item, creatureType + 'Name');
@@ -34,7 +50,7 @@ async function use({workflow}) {
                         value: hpValue
                     },
                     ac: {
-                        flat: 13 + workflow.actor.system.attributes.prof
+                        flat: 13 + workflow.actor.system.abilities.wis.mod
                     }
                 }
             },
@@ -42,7 +58,7 @@ async function use({workflow}) {
                 name,
                 disposition: workflow.token.document.disposition
             },
-            items: [primalBondFeatureData, dodgeData]
+            items: itemsToAdd
         },
         token: {
             name,
@@ -57,42 +73,61 @@ async function use({workflow}) {
         genericUtils.setProperty(updates, 'token.texture.src', tokenImg);
     }
     if (creatureType === 'land') {
-        let chargeData = await Summons.getSummonItem('Charge', {}, workflow.item, {flatDC: true, translate: 'CHRISPREMADES.CommonFeatures.Charge', identifier: 'primalCompanionCharge'});
-        let maulData = await Summons.getSummonItem('Maul', {}, workflow.item, {translate: 'CHRISPREMADES.CommonFeatures.Maul', identifier: 'primalCompanionMaul', flatAttack: true});
-        if (!chargeData || !maulData) {
+        let selection = await dialogUtils.buttonDialog(workflow.item.name, 'CHRISPREMADES.Dialog.DamageType', [
+            ['DND5E.DamageBludgeoning', 'bludgeoning'],
+            ['DND5E.DamagePiercing', 'piercing'],
+            ['DND5E.DamageSlashing', 'slashing']
+        ]);
+        if (!selection) selection = 'slashing';
+        let types = [selection];
+        if (exceptionalTraining) {
+            types.push('force');
+        }
+        let beastsStrikeData = await Summons.getSummonItem('Beast\'s Strike (Land)', {}, workflow.item, {flatAttack: true, translate: 'CHRISPREMADES.Macros.PrimalCompanion.BeastsStrike', identifier: 'primalCompanionLandBeastsStrike', rules: 'modern'});
+        if (!beastsStrikeData) {
             errors.missingPackItem();
             return;
         }
-        if (classLevel >= 7) {
-            chargeData.system.properties.push('mgc');
-            maulData.system.properties.push('mgc');
-        }
-        updates.actor.items.push(chargeData, maulData);
+        let attackActivity = Object.entries(beastsStrikeData.system.activities).map(a => a[1]).find(a => a.type === 'attack');
+        attackActivity.damage.parts[0].types = new Set(types);
+        updates.actor.items.push(beastsStrikeData);
         genericUtils.setProperty(updates, 'actor.system.attributes.movement', {walk: 40, climb: 40});
     } else if (creatureType === 'sea') {
-        let amphibiousData = await Summons.getSummonItem('Amphibious', {}, workflow.item, {translate: 'CHRISPREMADES.CommonFeatures.Amphibious', identifier: 'primalCompanionAmphibious'});
-        let bindingStrikeData = await Summons.getSummonItem('Binding Strike', {}, workflow.item, {flatAttack: true, flatDC: true, translate: 'CHRISPREMADES.Macros.PrimalCompanion.BindingStrike', identifier: 'primalCompanionBindingStrike'});
-        if (!amphibiousData || !bindingStrikeData) {
+        let amphibiousData = await Summons.getSummonItem('Amphibious', {}, workflow.item, {translate: 'CHRISPREMADES.CommonFeatures.Amphibious', identifier: 'primalCompanionAmphibious', rules: 'modern'});
+        let selection = await dialogUtils.buttonDialog(workflow.item.name, 'CHRISPREMADES.Dialog.DamageType', [
+            ['DND5E.DamageBludgeoning', 'bludgeoning'],
+            ['DND5E.DamagePiercing', 'piercing'],
+        ]);
+        if (!selection) selection = 'bludgeoning';
+        let types = [selection];
+        if (exceptionalTraining) {
+            types.push('force');
+        }
+        let beastsStrikeData = await Summons.getSummonItem('Beast\'s Strike (Sea)', {}, workflow.item, {flatAttack: true, translate: 'CHRISPREMADES.Macros.PrimalCompanion.BeastsStrike', identifier: 'primalCompanionSeaBeastsStrike', rules: 'modern'});
+        if (!amphibiousData || !beastsStrikeData) {
             errors.missingPackItem();
             return;
         }
-        if (classLevel >= 7) {
-            bindingStrikeData.system.properties.push('mgc');
-        }
-        updates.actor.items.push(amphibiousData, bindingStrikeData);
-        genericUtils.setProperty(updates, 'actor.system.attributes.movement', {walk: genericUtils.convertDistance(5), swim: genericUtils.convertDistance(60)});
+        let attackActivity = Object.entries(beastsStrikeData.system.activities).map(a => a[1]).find(a => a.type === 'attack');
+        attackActivity.damage.parts[0].types = new Set(types);
+        beastsStrikeData.flags['chris-premades'].config.generic.autoGrapple.dc = workflow.actor.system.attributes.spell.dc;
+        updates.actor.items.push(amphibiousData, beastsStrikeData);
+        genericUtils.setProperty(updates, 'actor.system.attributes.movement', {walk: 5, swim: 60});
     } else {
         hpValue = 4 + 4 * classLevel;
-        let flybyData = await Summons.getSummonItem('Flyby', {}, workflow.item, {translate: 'CHRISPREMADES.CommonFeatures.Flyby', identifier: 'primalCompanionFlyby'});
-        let shredData = await Summons.getSummonItem('Shred', {}, workflow.item, {translate: 'CHRISPREMADES.CommonFeatures.Shred', identifier: 'primalCompanionShred', flatAttack: true});
-        if (!flybyData || !shredData) {
+        let beastsStrikeData = await Summons.getSummonItem('Beast\s Strike (Sky)', {}, workflow.item, {flatAttack: true, translate: 'CHRISPREMADES.Macros.PrimalCompanion.BeastsStrike', identifier: 'primalCompanionSkyBeastsStrike', rules: 'modern'});
+        if (!flybyData || !beastsStrikeData) {
             errors.missingPackItem();
             return;
         }
-        if (classLevel >= 7) {
-            shredData.system.properties.push('mgc');
+        let types = ['slashing'];
+        if (exceptionalTraining) {
+            types.push('force');
         }
-        updates.actor.items.push(flybyData, shredData);
+        let attackActivity = Object.entries(beastsStrikeData.system.activities).map(a => a[1]).find(a => a.type === 'attack');
+        attackActivity.damage.parts[0].types = new Set(types);
+        let flybyData = await Summons.getSummonItem('Flyby', {}, workflow.item, {translate: 'CHRISPREMADES.CommonFeatures.Flyby', identifier: 'primalCompanionFlyby', rules: 'modern'});
+        updates.actor.items.push(beastsStrikeData, flybyData);
         genericUtils.mergeObject(updates, {
             actor: {
                 system: {
@@ -138,7 +173,7 @@ async function use({workflow}) {
         });
     }
     let animation = itemUtils.getConfig(workflow.item, creatureType + 'Animation') ?? 'none';
-    let identifiersToVae = ['primalCompanionDodge', 'primalCompanionCharge', 'primalCompanionMaul', 'primalCompanionBindingStrike', 'primalCompanionShred'];
+    let identifiersToVae = ['primalCompanionDodge', 'primalCompanionBeastsStrikeLand', 'primalCompanionBeastsStrikeSea', 'primalCompanionBeastsStrikeSky'];
     await Summons.spawn(sourceActor, updates, workflow.item, workflow.token, {
         range: 10,
         animation,
@@ -160,21 +195,25 @@ async function use({workflow}) {
         }
     });
 }
-async function earlySea({workflow}) {
-    let selection = await dialogUtils.buttonDialog(workflow.item.name, 'CHRISPREMADES.Dialog.DamageType', [
-        ['DND5E.DamagePiercing', 'piercing'],
-        ['DND5E.DamageBludgeoning', 'bludgeoning']
-    ]);
-    if (!selection) selection = 'piercing';
-    let activityData = activityUtils.withChangedDamage(workflow.activity, {}, [selection]);
-    workflow.item = itemUtils.cloneItem(workflow.item, {
-        ['system.activities.' + workflow.activity.id]: activityData
-    });
-    workflow.activity = workflow.item.system.activities.get(workflow.activity.id);
+async function damage({workflow}) {
+    let ownerActor = await fromUuid(workflow.actor.flags['chris-premades'].summons.control.actor);
+    let bestialFury = itemUtils.getItemByIdentifier(ownerActor, 'bestialFury');
+    if (!bestialFury) return;
+    if (workflow.hitTargets.size !== 1) return;
+    if (!workflowUtils.isAttackType(workflow, 'attack')) return;
+    let effect = effectUtils.getEffectByIdentifier(ownerActor, 'huntersMark');
+    if (!effect) return;
+    let {targets: validTargetUuids, formula, damageType} = effect.flags['chris-premades'].huntersMark;
+    if (!validTargetUuids.includes(workflow.hitTargets.first().document.uuid)) return;
+    let item = workflow.item;
+    if (!combatUtils.perTurnCheck(item, 'bestialFury', false, workflow.token.id)) return;
+    await workflowUtils.bonusDamage(workflow, formula, {damageType});
+    await combatUtils.setTurnCheck(item, 'bestialFury');
 }
 export let primalCompanion = {
     name: 'Primal Companion',
-    version: '1.1.0',
+    version: '1.3.79',
+    rules: 'modern',
     hasAnimation: true,
     midi: {
         item: [
@@ -288,15 +327,16 @@ export let primalCompanion = {
         },
     ]
 };
-export let primalCompanionSea = {
-    name: 'Primal Companion: Sea',
+export let bestialFury = {
+    name: 'Bestial Fury',
     version: primalCompanion.version,
+    rules: primalCompanion.rules,
     midi: {
         item: [
             {
-                pass: 'preambleComplete',
-                macro: earlySea,
-                priority: 50
+                pass: 'damageRollComplete',
+                macro: damage,
+                priority: 250
             }
         ]
     }