+++ scripts/macros/2024/classFeatures/ranger/foeSlayer.js
@@ -1,0 +1,24 @@
+import {constants} from '../../../../utils.js';
+export let foeSlayer = {
+    name: 'Foe Slayer',
+    version: '1.3.78',
+    config: [
+        {
+            value: 'formula',
+            label: 'CHRISPREMADES.Config.Formula',
+            type: 'text',
+            default: '1d10',
+            homebrew: true,
+            category: 'homebrew'
+        },
+        {
+            value: 'damageType',
+            label: 'CHRISPREMADES.Config.DamageType',
+            type: 'select',
+            default: 'force',
+            options: constants.damageTypeOptions,
+            homebrew: true,
+            category: 'homebrew'
+        },
+    ]
+};